requirements.txt
aiohttp==3.6.2
amqp==2.5.2
APScheduler==3.6.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
rsa==4.0
Safety-chan==1.0
six==1.14.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
